feature: check if user is owner of an object

+++ app/services/favourite_service.py
@@ -6,6 +6,7 @@ from fastapi import HTTPException
 from repositories.favourite_repository import FavouriteRepository
 from schemas.favourite import FavouriteInput, FavouriteOfferOutput
 from repositories.offer_repository import OfferRepository
+from services.user_service import UserService
 
 
 class FavouriteService:
@@ -13,6 +14,7 @@ class FavouriteService:
     def __init__(self, session: Session):
         self.repository = FavouriteRepository(session)
         self.repository_offer = OfferRepository(session)
+        self.user_service = UserService(session)
 
     def create(self, data: FavouriteInput) -> FavouriteInput:
         if self.repository.offer_saved_by_user(data.user_id, data.offer_id):
@@ -24,11 +26,14 @@ class FavouriteService:
         favourite = self.repository.create(data)
         return FavouriteInput(**favourite.__dict__)
 
-    def delete(self, _id: UUID4) -> bool:
+    def delete(self, _id: UUID4, user_id: UUID4) -> bool:
         if not self.repository.favourite_exists_by_id(_id):
             raise HTTPException(status_code=404, detail="Favourite not found")
 
         favourite = self.repository.get_favourite(_id)
+
+        if favourite.user_id != user_id:
+            raise HTTPException(status_code=401, detail="Unauthorized")
 
         return self.repository.delete(favourite)
